+++ chaosk8s/__init__.py
@@ -11,7 +11,7 @@ from logzero import logger
 
 
 __all__ = ["create_k8s_api_client", "discover", "__version__"]
-__version__ = '0.24.1'
+__version__ = '0.25.0'
 
 
 def has_local_config_file():